file sharing

+++ app/app/home/_components/delete-file-button.tsx
@@ -3,15 +3,23 @@
 import DeleteIcon from '@mui/icons-material/Delete';
 import { useRouter } from "next/navigation"
 
-export default function DeleteFileButton(props: { FILE_ID: string }): JSX.Element {
+export default function DeleteFileButton(props: { FILE_ID: string, IS_OWNER: boolean }): JSX.Element {
 	const router = useRouter()
 
 	const deleteFile = () => {
-		fetch(`/api/files/${props.FILE_ID}`, {
-			method: "DELETE"
-		}).then(() => {
-			router.refresh()
-		})
+		if(!props.IS_OWNER) {
+			fetch(`/api/files/${props.FILE_ID}/unshare`, {
+				method: "POST",
+			}).then((resp) => {
+				router.refresh()
+			})
+		} else {
+			fetch(`/api/files/${props.FILE_ID}`, {
+				method: "DELETE"
+			}).then(() => {
+				router.refresh()
+			})
+		}
 	}
 
 	return (
